Moved database models to Models\Database folder
