+++ PS2 Assistant/Models/Database/Roles.cs
@@ -1,5 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations;
-
+﻿namespace PS2_Assistant.Models.Database;
 public class Roles
 {
     public int Id { get; set; }
